+++ ChangeLog.markdown
@@ -6,6 +6,7 @@ This is the list of changes for the FinderFacade 1.0 release series.
 FinderFacade 1.0.1
 ------------------
 
+* Added `<include>` element as parent for the `<directory>` and `<file>` elements.
 * Renamed `<pattern>` to `<name>` as `Finder::name()` accepts globs, strings, or regular expressions.
 
 FinderFacade 1.0.0